# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Platform Wallet Point Gamifikasi Kampus - Backend API",
        "version": "1.0",
        "title": "Wallet Point API",
        "contact": {
            "name": "API Support",
            "email": "support@campus.edu"
        }
    },
    "host": "localhost:8102",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter your bearer token in the format: Bearer <token>"
        }
    },
    "paths": {
...
        "/mahasiswa/wallet": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Mahasiswa - Wallet"
                ],
                "summary": "Get My Wallet Balance",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/mahasiswa/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Mahasiswa - Wallet"
                ],
                "summary": "Get My Transaction History",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "default": 1
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "type": "integer",
                        "default": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/admin/wallets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Admin - Wallets"
                ],
                "summary": "Get All Wallets",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/admin/wallet/adjustment": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Admin - Wallets"
                ],
                "summary": "Adjust User Balance",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "user_id": {
                                    "type": "integer"
                                },
                                "amount": {
                                    "type": "number"
                                },
                                "type": {
                                    "type": "string",
                                    "enum": [
                                        "credit",
                                        "debit",
                                        "adjustment"
                                    ]
                                },
                                "description": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Adjustment successful"
                    }
                }
            }
        },
...
    }
}